[IN-302] refactor: refactor messenger components
