+++ src/features/messenger/model/usePartnerInfo.ts
@@ -3,9 +3,9 @@ import avatarPlaceholder from '@/assets/images/user-avatar-placeholder.jpg'
 import { messengerStore } from './stores/messengerStore'
 
 export const usePartnerInfo = () => {
+  const dialogPartnerInfo = messengerStore.dialogPartnerInfo
   const isChatLoading = messengerStore.isChatLoading
 
-  const dialogPartnerInfo = messengerStore.dialogPartnerInfo
   const avatar =
     dialogPartnerInfo && dialogPartnerInfo.avatars.length !== 0
       ? dialogPartnerInfo.avatars[1].url
