+++ src/features/messenger/ui/FindChat.tsx
@@ -7,19 +7,13 @@ export const FindChat = observer(() => {
   const { inputText, onFindChat, searchPlaceholder } = useFindChat()
 
   return (
-    <div
-      className={
-        'px-3 col-span-1 row-span-1 border-r border-b border-dark-300 bg-dark-500 flex items-center'
-      }
-    >
-      <TextField
-        bottomMarginForError={false}
-        label={''}
-        onChange={onFindChat}
-        placeholder={searchPlaceholder}
-        type={'search'}
-        value={inputText}
-      />
-    </div>
+    <TextField
+      bottomMarginForError={false}
+      label={''}
+      onChange={onFindChat}
+      placeholder={searchPlaceholder}
+      type={'search'}
+      value={inputText}
+    />
   )
 })
